--- a/lib/image/image.go
+++ b/lib/image/image.go
@@ -13,33 +13,34 @@ import (
 	"math"
 	"sync"
 
-	"github.com/BurntSushi/graphics-go/graphics"
-	"github.com/BurntSushi/graphics-go/graphics/interp"
 	"github.com/Kagami/go-avif"
 	"github.com/chai2010/webp"
+	"github.com/disintegration/gift"
 	"github.com/ieee0824/libcmyk"
 	imgproxyerr "github.com/livesense-inc/fanlin/lib/error"
-	"github.com/nfnt/resize"
 	"github.com/rwcarlsen/goexif/exif"
 	_ "golang.org/x/image/bmp"
 )
 
-var affines map[int]graphics.Affine = map[int]graphics.Affine{
-	1: graphics.I,
-	2: graphics.I.Scale(-1, 1),
-	3: graphics.I.Scale(-1, -1),
-	4: graphics.I.Scale(1, -1),
-	5: graphics.I.Rotate(toRadian(90)).Scale(-1, 1),
-	6: graphics.I.Rotate(toRadian(90)),
-	7: graphics.I.Rotate(toRadian(-90)).Scale(-1, 1),
-	8: graphics.I.Rotate(toRadian(-90)),
+var affines = map[int]gift.Filter{
+	2: gift.FlipHorizontal(),
+	3: gift.Rotate180(),
+	4: gift.FlipVertical(),
+	5: gift.Transpose(),
+	6: gift.Rotate270(),
+	7: gift.Transverse(),
+	8: gift.Rotate90(),
 }
 
 var mlConverterCache = &sync.Map{}
 
 type Image struct {
-	img    image.Image
-	format string
+	img         image.Image
+	format      string
+	orientation int
+	fillColor   color.Color
+	outerBounds image.Rectangle
+	filter      *gift.GIFT
 }
 
 func (i *Image) ConvertColor(networkPath string) error {
@@ -188,118 +189,66 @@ func EncodeAVIF(buf io.Writer, img *image.Image, q int) error {
 	return nil
 }
 
-// DecodeImage is return image.Image
 func DecodeImage(r io.Reader) (*Image, error) {
-	img, format, err := decode(r)
-	return &Image{img: img, format: format}, imgproxyerr.New(imgproxyerr.WARNING, err)
+	img, format, orientation, err := decode(r)
+	wrapper := Image{
+		img:         img,
+		format:      format,
+		orientation: orientation,
+		outerBounds: img.Bounds(),
+		filter:      gift.New(),
+	}
+	return &wrapper, imgproxyerr.New(imgproxyerr.WARNING, err)
 }
 
-// アス比を維持した時の長さを取得する
-func keepAspect(img image.Image, w uint, h uint) (uint, uint) {
-	r := img.Bounds()
-	if int(w)*r.Max.Y < int(h)*r.Max.X {
-		return w, 0
+func (i *Image) Process() {
+	bounds := i.filter.Bounds(i.img.Bounds())
+	dest := image.NewRGBA(bounds)
+	i.filter.Draw(dest, i.img)
+
+	if dest.Bounds() == i.outerBounds || i.fillColor == nil {
+		i.img = dest
+		return
+	}
+
+	bg := image.NewRGBA(i.outerBounds)
+	draw.Draw(bg, bg.Bounds(), &image.Uniform{i.fillColor}, image.Point{}, draw.Src)
+	centerH := math.Abs(float64(i.outerBounds.Max.Y-bounds.Max.Y)) / 2.0
+	centerW := math.Abs(float64(i.outerBounds.Max.X-bounds.Max.X)) / 2.0
+	center := bounds.Min.Sub(image.Pt(int(centerW), int(centerH)))
+	draw.Draw(bg, bg.Bounds(), dest, center, draw.Over)
+	i.img = bg
+}
+
+func (i *Image) ResizeAndFill(w, h uint, c color.Color) {
+	if w == 0 || h == 0 {
+		return
+	}
+	innerW := w
+	innerH := h
+	r := i.img.Bounds()
+	if int(innerW)*r.Max.Y < int(innerH)*r.Max.X {
+		innerH = 0
 	} else {
-		return 0, h
+		innerW = 0
 	}
+	i.filter.Add(gift.Resize(int(innerW), int(innerH), gift.LanczosResampling))
+	i.outerBounds = image.Rect(0, 0, int(w), int(h))
+	i.fillColor = c
 }
 
-func resizeImage(img image.Image, w uint, h uint, maxWidth uint, maxHeight uint) image.Image {
-	if img == nil {
-		return nil
-	}
-	//大きすぎる値はサポートしない
-	w = max(w, maxWidth)
-	h = max(h, maxHeight)
-	w, h = keepAspect(img, w, h)
-	// 速度・負荷的な問題出た時はアルゴリズム変更
-	return resize.Resize(w, h, img, resize.Lanczos3)
-}
-
-func resizeAndFillImage(img image.Image, w uint, h uint, c color.Color, maxWidth uint, maxHeight uint) image.Image {
-	if img == nil {
-		return nil
-	}
-	if maxWidth < w || maxHeight < h {
-		return img
-	}
-	ch0 := make(chan image.Image)
-	ch1 := make(chan *image.RGBA)
-
-	// ココらへんの並列化はベンチマーク次第で変更する
-	go func() {
-		ch0 <- resizeImage(img, w, h, maxWidth, maxHeight)
-	}()
-	go func() {
-		ch1 <- image.NewRGBA(image.Rect(0, 0, int(w), int(h)))
-	}()
-	resizedImage := <-ch0
-	m := <-ch1
-
-	zeroPoint := image.Point{}
-	draw.Draw(m, m.Bounds(), &image.Uniform{c}, zeroPoint, draw.Src)
-
-	//画像の中心座標を計算
-	centerH := int(h)/2 - (resizedImage.Bounds().Max.Y / 2)
-	centerW := int(w)/2 - (resizedImage.Bounds().Max.X / 2)
-
-	if resizedImage.Bounds().Max.X == int(w) {
-		draw.Draw(m, m.Bounds(), resizedImage, resizedImage.Bounds().Min.Sub(image.Pt(0, centerH)), draw.Over)
-	} else if resizedImage.Bounds().Max.Y == int(h) {
-		draw.Draw(m, m.Bounds(), resizedImage, resizedImage.Bounds().Min.Sub(image.Pt(centerW, 0)), draw.Over)
-	} else {
-		return resizedImage
-	}
-	return m
-}
-
-func (i *Image) ResizeAndFill(w uint, h uint, c color.Color, maxW uint, maxH uint) {
-	if maxW < w || maxH < h {
+func (i *Image) Crop(w, h uint) {
+	if w == 0 || h == 0 {
 		return
 	}
-	if h == 0 || w == 0 {
-		return
-	}
-	if c == nil {
-		i.img = resizeImage(i.img, w, h, maxW, maxH)
-		return
-	}
-	i.img = resizeAndFillImage(i.img, w, h, c, maxW, maxH)
+	i.filter.Add(gift.ResizeToFill(int(w), int(h), gift.LanczosResampling, gift.CenterAnchor))
+	i.outerBounds = image.Rect(0, 0, int(w), int(h))
 }
 
-func crop(img image.Image, w uint, h uint) image.Image {
-	if img == nil {
-		return nil
+func (i *Image) ApplyOrientation() {
+	if affine, ok := affines[i.orientation]; ok {
+		i.filter.Add(affine)
 	}
-	if h == 0 || w == 0 {
-		return img
-	}
-
-	orgW := img.Bounds().Max.X
-	orgH := img.Bounds().Max.Y
-
-	r := float64(orgW) / float64(w)
-	if (float64(orgW) / float64(orgH)) > (float64(w) / float64(h)) {
-		r = float64(orgH) / float64(h)
-	}
-
-	startW := orgW/2 - int(float64(w)*r/2)
-	startH := orgH/2 - int(float64(h)*r/2)
-
-	result := image.NewRGBA(image.Rect(0, 0, int(float64(w)*r), int(float64(h)*r)))
-
-	for y := 0; y < int(float64(h)*r); y++ {
-		for x := 0; x < int(float64(w)*r); x++ {
-			c := img.At(x+startW, y+startH)
-			result.Set(x, y, c)
-		}
-	}
-
-	return result
-}
-
-func (i *Image) Crop(w uint, h uint) {
-	i.img = crop(i.img, w, h)
 }
 
 func (i *Image) GetImg() *image.Image {
@@ -310,36 +259,13 @@ func (i *Image) GetFormat() string {
 	return i.format
 }
 
-func Set404Image(buf io.Writer, data io.Reader, w uint, h uint, c color.Color, maxW uint, maxH uint) error {
+func Set404Image(buf io.Writer, data io.Reader, w uint, h uint, c color.Color) error {
 	img, err := DecodeImage(data)
 	if err != nil {
 		return imgproxyerr.New(imgproxyerr.ERROR, err)
 	}
-	img.ResizeAndFill(w, h, c, maxW, maxH)
+	img.ResizeAndFill(w, h, c)
 	return EncodeJpeg(buf, img.GetImg(), jpeg.DefaultQuality)
-}
-
-func toRadian(n int) float64 {
-	return float64(n) * math.Pi / 180.0
-}
-
-func applyOrientation(s image.Image, o int) (d draw.Image, e error) {
-	bounds := s.Bounds()
-	if o == 0 {
-		o = 1
-	}
-	if o >= 5 && o <= 8 {
-		bounds = rotateRect(bounds)
-	}
-	d = image.NewRGBA(bounds)
-	affine := affines[o]
-	e = affine.TransformCenter(d, s, interp.Bilinear)
-	return
-}
-
-func rotateRect(r image.Rectangle) image.Rectangle {
-	s := r.Size()
-	return image.Rectangle{r.Min, image.Point{s.Y, s.X}}
 }
 
 func readOrientation(r io.Reader) (o int, err error) {
@@ -358,19 +284,15 @@ func readOrientation(r io.Reader) (o int, err error) {
 	return
 }
 
-func decode(r io.Reader) (d image.Image, format string, err error) {
+func decode(r io.Reader) (d image.Image, format string, o int, err error) {
 	var buf bytes.Buffer
 	tee := io.TeeReader(r, &buf)
 
-	s, format, err := image.Decode(tee)
+	d, format, err = image.Decode(tee)
 	if err != nil {
 		return
 	}
 
-	o, err := readOrientation(&buf)
-	if err != nil {
-		return s, format, nil
-	}
-	d, err = applyOrientation(s, o)
+	o, _ = readOrientation(&buf)
 	return
 }