## Problem
The following process is too slow. It spends lots of cpu time and memory allocation. It looks like a bottleneck.
https://github.com/livesense-inc/fanlin/blob/45a1655871dda0fa9b19840888af028705d93b78/lib/image/image.go#L336

The above process depends on the following library. Unfortunately, the last update is a decade ago.
https://github.com/BurntSushi/graphics-go

CPU
```
      flat  flat%   sum%        cum   cum%
     280ms 17.95% 17.95%      280ms 17.95%  github.com/nfnt/resize.resizeRGBA
     220ms 14.10% 32.05%      730ms 46.79%  github.com/BurntSushi/graphics-go/graphics/interp.bilinearGeneral
     150ms  9.62% 41.67%      230ms 14.74%  runtime.mallocgc
      70ms  4.49% 46.15%       70ms  4.49%  image/color.YCbCr.RGBA
      70ms  4.49% 50.64%      300ms 19.23%  image/jpeg.(*decoder).processSOS
      70ms  4.49% 55.13%      110ms  7.05%  image/jpeg.(*decoder).reconstructBlock
      60ms  3.85% 58.97%       60ms  3.85%  github.com/BurntSushi/graphics-go/graphics/interp.findLinearSrc
      60ms  3.85% 62.82%       60ms  3.85%  runtime.deductAssistCredit
      50ms  3.21% 66.03%       60ms  3.85%  image/jpeg.(*decoder).receiveExtend
      50ms  3.21% 69.23%      330ms 21.15%  runtime.convTnoptr
```

Memory
```
      flat  flat%   sum%        cum   cum%
      83MB 33.11% 33.11%       83MB 33.11%  image.(*YCbCr).At
   81.50MB 32.52% 65.63%   164.50MB 65.63%  github.com/BurntSushi/graphics-go/graphics/interp.bilinearGeneral
   39.15MB 15.62% 81.25%    39.15MB 15.62%  bytes.growSlice
   27.04MB 10.79% 92.03%    27.04MB 10.79%  image.NewRGBA
   15.45MB  6.16% 98.20%    15.45MB  6.16%  image.NewYCbCr
    1.51MB   0.6% 98.80%     1.51MB   0.6%  io.ReadAll
         0     0% 98.80%    23.02MB  9.19%  bufio.(*Reader).Read
         0     0% 98.80%    16.12MB  6.43%  bytes.(*Buffer).ReadFrom
         0     0% 98.80%    23.02MB  9.19%  bytes.(*Buffer).Write
         0     0% 98.80%    39.15MB 15.62%  bytes.(*Buffer).grow
```

## Solution
I found an awesome library which is named "gift". It is also used in [hugo](https://github.com/gohugoio/hugo) project.
https://github.com/disintegration/gift

So I think we need to migrate some libraries to the gift for the performance improvement.

* https://github.com/BurntSushi/graphics-go
  * The last update is a decade ago.
* https://github.com/nfnt/resize
  * repository archived
  * We can use Lanczos algorithm for interpolation by the gift as well.

Several operations for our image processing can be done by only the gift.

* applying orientation
* cropping
* resize

The gift can do processing in parallel as default. We can pass several filters to it, then we can draw image at last.

## Benchmark
### Before
```
$ echo 'GET http://127.0.0.1:3000/baz/Lenna.jpg?w=300&h=200&rgb=32,32,32' | vegeta attack -header='user-agent: vegeta' -rate=50 -duration=180s | tee results.bin | vegeta report
Requests      [total, rate, throughput]         9000, 50.01, 49.98
Duration      [total, attack, wait]             3m0s, 3m0s, 88.627ms
Latencies     [min, mean, 50, 90, 95, 99, max]  87.575ms, 94.242ms, 93.583ms, 98.539ms, 100.593ms, 105.215ms, 116.145ms
Bytes In      [total, mean]                     87327000, 9703.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:9000
Error Set:
```

### After
```
$ echo 'GET http://127.0.0.1:3000/baz/Lenna.jpg?w=300&h=200&rgb=32,32,32' | vegeta attack -header='user-agent: vegeta' -rate=50 -duration=180s | tee results.bin | vegeta report
Requests      [total, rate, throughput]         9000, 50.01, 50.00
Duration      [total, attack, wait]             3m0s, 3m0s, 29.395ms
Latencies     [min, mean, 50, 90, 95, 99, max]  28.803ms, 30.883ms, 30.859ms, 31.896ms, 32.219ms, 32.924ms, 43.37ms
Bytes In      [total, mean]                     87183000, 9687.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:9000
Error Set:
```

![3_times_faster](https://github.com/user-attachments/assets/c8c25063-847d-40ba-9935-7b4fbd49da44)
(c) Bandai Namco Holdings, Sotsu

## Related links
* https://pkg.go.dev/github.com/disintegration/gift
* https://pkg.go.dev/image
* #91
* #56